requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.49
langchain-openai==0.3.11
langsmith==0.3.19
openai==1.68.2
orjson==3.10.16
packaging==24.2
pydantic==2.10.6
pydantic_core==2.27.2
python-decouple==3.8
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typing_extensions==4.13.0
urllib3==2.3.0
zstandard==0.23.0
